requirements.txt
tensorflow
gradio
numpy
scipy
pillow